# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Роботоподбор API
  version: 0.1.0
  description: >
    Стаб контракта для фронтенда. Бизнес-логика расчёта живёт на бэкенде,
    фронт только отображает готовые значения.
  license:
    name: Proprietary
servers:
  - url: /api
security:
  - sessionCookie: []
paths:
...
  /selection:
    post:
      operationId: selectSolutions
      summary: Подбор решений под паспорт объекта
      security: []
      description: >
        Подбор решений под паспорт объекта (ТЗ 3.4): для каждого решения —
        статус (рекомендовано / требует проверки / исключено), причины
        соответствия, ограничения, недостающие данные и разложение балла по критериям.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SelectionRequest' }
      responses:
        '200':
          description: Результат подбора
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelectionResult' }
        '400':
          description: Некорректный запрос
  /object-types/{slug}/parameters/template:
    get:
      operationId: getParametersTemplate
      summary: Шаблон паспорта объекта (xlsx/csv)
      security: []
      description: Шаблон паспорта объекта для заполнения и загрузки (ТЗ 3.2.3).
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: format
          in: query
          required: false
          schema: { type: string, enum: [xlsx, csv], default: xlsx }
      responses:
        '200':
          description: Файл шаблона
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
            text/csv:
              schema: { type: string }
        '404':
          description: Тип объекта не найден
...
  /catalog/solutions:
    get:
      operationId: listSolutions
      summary: Каталог решений
      security: []
      parameters:
        - name: objectType
          in: query
          description: Оставить только решения, применимые к типу объекта.
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Каталог решений
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Solution' }
    post:
      operationId: createSolution
      summary: Добавить решение в каталог
      description: Добавление решения в каталог. Доступно роли admin.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Solution' }
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Solution' }
        '400':
          description: Некорректные данные
        '403':
          description: Недостаточно прав
...
  /object-types/{slug}/parameters:
    get:
      operationId: getObjectParameters
      summary: Состав паспорта объекта
      security: []
      description: Состав полей паспорта объекта зависит от типа объекта.
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Схема полей
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ParameterField' }
  /object-types/{slug}/parameters/import:
    post:
      operationId: importObjectParameters
      summary: Загрузить паспорт объекта из файла
      security: []
      description: Разбор загруженного Excel/CSV в значения паспорта объекта.
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Распознанные значения
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportedParameters' }
        '422':
          description: Файл не распознан
...
  /catalog/solutions/{solutionId}:
    put:
      operationId: updateSolution
      summary: Изменить решение
      description: Доступно роли admin.
      parameters:
        - name: solutionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Solution' }
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Solution' }
        '403':
          description: Недостаточно прав
    delete:
      operationId: deleteSolution
      summary: Удалить решение
      parameters:
        - name: solutionId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Удалено
        '403':
          description: Недостаточно прав
  /catalog/solutions/export:
    get:
      operationId: exportCatalog
      summary: Выгрузить каталог
      description: Выгрузка каталога в таблицу организатора (ТЗ 3.3.2). Только администратор.
      parameters:
        - name: format
          in: query
          required: false
          schema: { type: string, enum: [xlsx, csv], default: xlsx }
      responses:
        '200':
          description: Файл каталога
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
            text/csv:
              schema: { type: string }
        '401': { description: Требуется вход }
        '403': { description: Недостаточно прав }
  /catalog/solutions/import:
    post:
      operationId: importCatalog
      summary: Загрузить каталог из файла
      description: >
        Загрузка каталога из xlsx/csv (ТЗ 3.8.2). Строка обновляет позицию по id или по
        паре «наименование + производитель», иначе создаёт новую. Ошибки — по строкам.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Итог загрузки
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CatalogImportResult' }
        '401': { description: Требуется вход }
        '403': { description: Недостаточно прав }
        '422': { description: Файл не передан или не читается }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: httpOnly-cookie сессии, выдаётся при входе и регистрации.
  schemas:
...
    ParameterField:
      type: object
      required: [id, label, kind]
      properties:
        id: { type: string }
        label: { type: string }
        hint: { type: string }
        unit: { type: string }
        section: { type: string, description: 'Заголовок группы полей в форме ввода параметров объекта' }
        kind:
          type: string
          enum: [number, text, select]
        min: { type: number }
        max: { type: number }
        defaultValue: { type: string }
        required: { type: boolean, description: 'Обязательный параметр (true) или дополнительный.' }
        source: { type: string, description: Источник значения по умолчанию (норматива). }
        options:
          type: array
          items:
            type: object
            required: [value, label]
            properties:
              value: { type: string }
              label: { type: string }
    ImportIssue:
      type: object
      required: [message]
      properties:
        row: { type: integer, description: Номер строки в файле (с 1). }
        fieldId: { type: string }
        label: { type: string }
        message: { type: string, description: Что не так и как исправить. }
    ImportedParameters:
      type: object
      required: [values, recognized]
      properties:
        values:
          type: object
          description: Только прошедшие проверку значения.
          additionalProperties: { type: string }
        recognized: { type: integer }
        skipped:
          type: array
          description: Строки, ключ которых не найден в паспорте.
          items: { type: string }
        errors:
          type: array
          description: Значения, не прошедшие проверку типа, единиц или диапазона.
          items: { $ref: '#/components/schemas/ImportIssue' }
        warnings:
          type: array
          description: Незаполненные обязательные поля — подставлено значение по умолчанию.
          items: { $ref: '#/components/schemas/ImportIssue' }
...
    CatalogImportResult:
      type: object
      required: [created, updated, errors]
      properties:
        created: { type: integer }
        updated: { type: integer }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ImportIssue' }
...
    DataConfidence:
      type: string
      enum: [confirmed, needs-review]
    Maturity:
      type: string
      enum: [operation, piloting, rnd]
...
    Solution:
      type: object
      description: >
        Карточка решения каталога. Группы полей повторяют таблицу 3.3 ТЗ:
        идентификация, технические характеристики, инфраструктура, экономика,
        применимость, качество данных. Денежные значения — млн ₽ с НДС.
      required: [id, name, vendor, useCase, price, payload, speed, maturity, confidence]
      properties:
        id: { type: string }
        name: { type: string }
        vendor: { type: string, description: Производитель. }
        useCase: { type: string, description: Назначение. }
        price: { type: string, description: 'Ориентировочная цена единицы, млн ₽.' }
        payload: { type: string, description: Грузоподъёмность для отображения. }
        speed: { type: string }
        maturity: { $ref: '#/components/schemas/Maturity' }
        confidence: { $ref: '#/components/schemas/DataConfidence' }
        source: { type: string, description: Название источника характеристик. }
        sourceDate: { type: string, description: Дата получения или обновления данных. }
        sourceUrl: { type: string, description: Ссылка на источник. }
        objectTypes:
          type: array
          description: Типы объектов, для которых решение применимо.
          items: { type: string }
        score: { type: number, description: 'Итоговый балл ранжирования, 0..100.' }
        scoreFactors:
          type: array
          description: Вклад критериев в позицию решения.
          items: { $ref: '#/components/schemas/ScoreFactor' }
        solutionType:
          type: string
          description: 'Тип решения: ключ справочника SOLUTION_TYPES (amr, fmr, asrs, cleaner, courier, inventory, security, …).'
        country: { type: string, description: Страна происхождения. }
        availability: { type: string, enum: [available, on-order, pilot], description: Статус доступности. }
        payloadKg: { type: number, description: 'Грузоподъёмность, кг (0 — решение не перевозит грузы).' }
        weightKg: { type: number, description: 'Собственная масса, кг.' }
        dimensions: { type: string, description: 'Габариты, Д×Ш×В.' }
        throughput: { type: number, description: Производительность в единицах throughputUnit. }
        throughputUnit: { type: string }
        autonomyHours: { type: number, description: 'Автономность, ч. Для стационарных систем 24 — работа от сети.' }
        positioningAccuracyMm: { type: number, description: 'Точность позиционирования, мм.' }
        navigation: { type: string, description: Тип навигации. }
        operatingConditions: { type: string, description: Допустимые условия эксплуатации. }
        environment: { type: string, enum: [indoor, outdoor, both] }
        minTempC: { type: number }
        maxTempC: { type: number }
        noiseDb: { type: number, description: 'Уровень шума, дБА.' }
        minAisleWidthM: { type: number, description: 'Минимальная ширина прохода, м.' }
        maxFloorDeviationMm: { type: number, description: 'Допуск ровности пола, мм на 2 м.' }
        minCeilingHeightM: { type: number, description: 'Минимальная высота помещения, м.' }
        elevatorIntegration: { type: boolean, description: Умеет ли вызывать лифт. }
        airsideCertified: { type: boolean, description: Допуск к работе на перроне. }
        infrastructure:
          type: object
          description: Требования к инфраструктуре.
          properties:
            floor: { type: string, description: Покрытие и проходы. }
            charging: { type: string }
            connectivity: { type: string }
            integration: { type: string }
            service: { type: string }
        costs:
          type: object
          description: 'Ориентировочная стоимость, млн ₽ с НДС.'
          properties:
            equipment: { type: number }
            software: { type: number }
            implementation: { type: number }
            maintenancePerYear: { type: number }
        acquisitionModels:
          type: array
          items: { type: string, enum: [purchase, leasing, raas] }
        lifespanYears: { type: number, description: 'Срок службы, лет.' }
        processes:
          type: array
          description: Поддерживаемые процессы (идентификаторы уровня «процесс» иерархии).
          items: { type: string }
        limitations: { type: array, items: { type: string } }
        cases: { type: array, description: Реализованные кейсы., items: { type: string } }
        unconfirmedFields:
          type: array
          description: Характеристики, значения которых приняты как допущение.
          items: { type: string }
        completeness:
          type: integer
          description: 'Полнота карточки по обязательным характеристикам, %. Считается сервером.'
        catalogCategory: { type: string, description: 'Код категории каталога роботов (AM, FL, FC, …).' }
        trl: { type: integer, description: 'Уровень готовности технологии (УГТ), 1–9.' }
        photos:
          type: array
          description: Имена файлов фото в /robots_photo.
          items: { type: string }
        objectFit:
          type: object
          description: >
            Откуда известно, что решение подходит типу объекта: declared — объект указан
            в каталоге, example — пример решения для объекта в материалах организатора,
            inferred — выведено по сценариям применения.
          additionalProperties: { type: string, enum: [declared, example, inferred] }
        fieldSources:
          type: object
          description: >
            Происхождение характеристик (ТЗ 3.3.4): ключ — поле карточки (payloadKg,
            dimensions, infrastructure.charging, …). Считается из источников каталога,
            правка администратора заменяет запись на «правка администратора».
          additionalProperties: { $ref: '#/components/schemas/FieldProvenance' }
...
    FieldProvenance:
      type: object
      required: [confirmed, sources]
      properties:
        confirmed: { type: boolean, description: 'true — значение в источнике точное; false — «около», значение серии, пересчёт или вывод.' }
        note: { type: string }
        sources:
          type: array
          items:
            type: object
            required: [kind, title]
            properties:
              kind: { type: string, enum: [organizer, vendor, reseller, media, registry, team, admin] }
              title: { type: string }
              url: { type: string }
              date: { type: string, description: 'ДД.ММ.ГГГГ' }
    ScoreFactor:
      type: object
      required: [label, weight]
      properties:
        label: { type: string }
        weight: { type: number, description: Набранные баллы. }
        max: { type: number, description: Максимум баллов по критерию. }
    SelectionRequest:
      type: object
      required: [objectType]
      properties:
        objectType: { type: string }
        parameters:
          type: object
          description: Значения паспорта объекта; пропущенные берутся по умолчанию.
          additionalProperties: { type: string }
    SelectionItem:
      type: object
      required: [solutionId, status, score, factors, reasons, limitations, missing, blockers]
      properties:
        solutionId: { type: string }
        status:
          type: string
          enum: [recommended, needs-review, excluded]
        score: { type: number }
        factors:
          type: array
          items: { $ref: '#/components/schemas/ScoreFactor' }
        reasons: { type: array, description: Причины соответствия., items: { type: string } }
        limitations: { type: array, items: { type: string } }
        missing: { type: array, description: Недостающие данные., items: { type: string } }
        blockers: { type: array, description: Причины исключения., items: { type: string } }
    SelectionResult:
      type: object
      required: [objectType, rulesVersion, summary, items]
      properties:
        objectType: { type: string }
        rulesVersion: { type: string }
        summary:
          type: object
          required: [recommended, needsReview, excluded]
          properties:
            recommended: { type: integer }
            needsReview: { type: integer }
            excluded: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/SelectionItem' }
...